+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed race condition that duplicated lines in progress
+- Fixed race condition that duplicated lines in progress https://github.com/willmcgugan/rich/issues/1144
 
 ## [10.0.0] - 2021-03-27