+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.0.1] - 2021-03-30
+
+### Fixed
+
+- Fixed race condition that duplicated lines in progress https://github.com/willmcgugan/rich/issues/1144
+
 ## [10.0.0] - 2021-03-27
 
 ### Changed